# Enhance DoH Logic