## 🗣 Description ##
- Remove the hard-dependency on IPv4 for the DoH call
- Adjust the DoH logic so that DoH calls are not made when DoH is blocked

## 💭 Motivation and context ##
Closes #737. Additionally, while working #737, made the following realization: when DoH is not available, a lot of time will be wasted on DoH calls. For example, if DoH calls are blocked by a proxy, it could take several seconds before an exception is thrown. Then it will retry the query. And it will do this for each domain name. On our G5 testing tenant, that results in 24 DoH calls, so that adds up. Now imagine the latency that would create for a large organization with many more domains...a change was needed. The change: when it becomes evident that DoH will be needed, first test to see if DoH is available. If it's not, skip all subsequent attempts to make a DoH call.

## 🧪 Testing ##
Tested in the following environments:
- A VM with an IPv6 interface
- A VM with an IPv6 interface with a hard-coded exception (`x/0`) before the traditional DNS query to force DoH
- My work computer on a network where DoH is blocked by the proxy
- My work computer on a network where DoH is available
- My work computer with a hard-coded exception (`x/0`) before the traditional DNS query to force DoH

I performed the following timing tests on a testing tenant with 4 domains:

On this branch with DoH unavailable (blocked by proxy):
- 3 DoH calls needed
- First run: 5.50 seconds
- Subsequent runs: Average of 1.79 seconds

On main with DoH unavailable (blocked by proxy):
- 24 DoH calls needed
- First run: 9.42 seconds
- Subsequent runs: Average of 7.87 seconds

## ✅ Pre-approval checklist ##
- [x] This PR has an informative and human-readable title.
- [x] Changes are limited to a single goal - *eschew scope creep!*
- [x] *All* future TODOs are captured in issues, which are referenced
      in code comments.
- [x] All relevant type-of-change labels have been added.
- [x] I have read the [CONTRIBUTING](../blob/develop/CONTRIBUTING.md) document.
- [x] These code changes follow [cisagov code standards](https://github.com/cisagov/development-guide).
- [x] All relevant repo and/or project documentation has been updated
      to reflect the changes in this PR.
- [x] Tests have been added and/or modified to cover the changes in this PR.
- [x] All new and existing tests pass.

## ✅ Pre-merge checklist ##
- [ ] Revert dependencies to default branches.
- [ ] Finalize version.